locale and about test to options folder
