--- bigbluebutton-tests/playwright/options/settings/locales/settings.spec.js
@@ -1,12 +1,0 @@
-const { test } = require('@playwright/test');
-const { Language } = require('./language');
-
-test.describe.parallel('Settings', () => {
-    test(`Locales`, async ({ browser, page }) => {
-      test.slow();
-      const language = new Language(browser, page);
-      await language.init(true, true);
-      await language.test();
-    });
-});
-
